requirements.txt
cryptography
Pillow
django
django-browser-reload
python-docx
qrcode
opencv-python-headless
numpy